Fix: containers made of iron would rust away and not drop their contents.
Was noticed by a player that was repeatedly dropping an iron safe into
shallow water. The iron safe would rust away completely, but anything
inside of it disppeared along with the safe. This commit addresses this,
and sets up a condition so that any iron container (iron large boxes and
chests are possible) will drop its contents upon rusting away
completely.

Since iron safes are substantial, make it so they can still rust away
but that the odds of it happening on the next water_damage() call are
pretty low.