Fix: another case of 'placing steed onto map, mstate:0, on Dlvl:() ?'.
This panic would occur whenever a players steed was hit with a monsters
clobber attack. The steed is supposed to be knocked back one or two tile
spaces, but was not legitimately freed from having a rider so it stayed
in place, hence the panic. This commit addresses that - what will happen
now, the player will be knocked off of their steed, and then the steed
can properly be knocked backwards.

While working on this, I noticed that if a player riding steed was hit
with a knock-back attack, they would be knocked back without any issue,
but their steed would be knocked back with them. That makes no sense. So
now, the player is effectively knocked out of the saddle and is forcibly
dismounted from their steed.