Pseudodragons given a range of different sounds.
The pseudodragon was lacking sound feedback , as non-hostile monsters
that use MS_HISS don't give any sound feedback at all. Added a bit here
for the pseudodragon as they are primarily a pet. Fun fact, dragons and
cats share some of the same sounds for certain events and emotional
states.