Fix: prevent removal of special artifacts on bones level and the owner is present.
The problem - player encounters bones on a level where a special
artifact resides with its owner, and that artifact (good example, the
magic 8-ball and the oracle level) was in possession of the player when
they died and left bones. The artifact would have its special artifact
properies stripped whether the rightful owner was present or not. This
patch does two things: it makes sure that the artifact is retained in
its correct state on that particular level if the owner of said artifact
is not present, and if the owner is present, it rightfully goes back
into its inventory. Thanks qt for hashing this one out.

Also in this commit - added a couple more unique monsters to the list
that shouldn't appear in bones.

Co-authored-by: entrez <me@entrez.cc>