Slight changes to intrinsics gained via crowning for infidel role.
Demons are inherently immune to sickness, but this resistance wasn't
carrying over when an infidel would be crowned and would assume demonic
form. So guarntee sick resistance for crowned infidels. Another thing I
noticed while working on this, was that if the infidel started out as an
illithid, they still kept their psychic resistance in demonic form.
That's been taken away, as no creatures save illithids have the correct
brain structure for such a resistance.