Some minor fixes to the racial priests and priestesses commit.
Made it so the high priests on Astral Plane do not change races,
otherwise it would be rather easy for a blinded telepathic player to
figure out which altar to go for. Also ensure that the high
priest/priestess in the sanctum is actually referred to as a high
priest/priestess.