More tweaks to racial priests/priestesses.
The changes:

* Removed dwarves and hobbits, added giants. Why? Because Giants are
another one of the base templates in monst.c that isn't used for
anything other that a base for player creation, whereas dwarves and
hobbits are also their own distinct monster. This matters because...
* When a racial priest is created, it takes the attributes of the
monster it becomes, including all of the flags set in its template in
monst.c, meaning that they could then be genocided, which is bad. So
I've purposely removed G_GENO from all racial priests as they are
created. During testing, I noticed that if you tried to genocide a dwarf
priest, it didn't work (yay) but it didn't work on any other dwarves
either (boo). So for now, we're only using races that have stand-alone
templates that never spawn randomly on their own (humans, elves, giants,
centaurs, illithids, and orcs).
* Adjusted the gear template for all priests (racial or not) in
makemon.c - noticed awhile back that I had added quarterstaff as a
possible weapon a priest could spawn with, but if they did, they
wouldn't use it because of the small shield they were wearing. So that's
changed, and there's also a specific gear template for giant priests vs
all the rest.

Some of these changes are really hacky. There's probably a better way to
do this (and racial shopkeepers as well, which also need to not be
genocidable), but this works for the time being.