Use quotes for spoken text from artifact defenders
Artifact defenders' speech was formatted like this:

    Andy the Investigator says: Did you really think I would hand over
    Grayswandir so easily?

But similar constructions elsewhere in the game typically use
`verbalize` to surround the spoken text with quotation marks:

    The priestess of Anhur intones: "Pilgrim, you enter a sacred place!"

This commit just brings artifact defenders' speech in line with this
format.

Also includes some minor consolidation of duplicate code, as well as
tests for deafness in cases where something the hero hears is described
when the artifact defender appears.