True dwarf/gnome placeholder templates, bit of a racial shopkeeper revamp.
Whilst working on racial priests, I was having to do some hacky things
to prevent said priests from being genocided. Only a couple of the mnum
types (PM_ELF and PM_HUMAN) were safe from genocide, I'm not sure why I
didn't make the others the same way. So that's been addressed, and their
stats have been adjusted to match their placeholder status (this was
actually done for PM_ELF in 3.7.0). And now I don't have to do any
hackery to prevent racial priests from being genocided.

But, I really missed having dwarves and gnomes in the mix - there is no
placeholder template for those. Until now, that is. Dwarves and gnomes
that the player will encounter in-game will appear as 'mountain dwarf'
and 'rock gnome'. Their stats are exactly the same as before, the only
change is in name only. So now, PM_DWARF and PM_GNOME can act as a true
placeholder template for race/role purposes, and for including them back
into the fold for racial priests or shopkeepers (gnomes don't do clergy
work, they prefer shopkeeping and wizard-like things).

So now, we're able to do a proper racial shopkeeper revamp. Got rid of
the hobbit (not going to mess with this one) as well as the nymph types,
mind flayers are now simply illithids, and you'll no longer see a crazy
mix of different types of human/elven/dwarvish whatever. A bit less
diversity, but this ensures that shopkeepers stay off-limits to genocide
as originally intended in vanilla NetHack.