Add new wielding_artifact function
Artifact effects were applied inconsistently, in some cases only
activating when the artifact is wielded in the main hand (uwep) and in
others also activating when the artifact is wielded in the off-hand
(uswapwep).  For example, Excalibur would only prevent bribing a demon
prince if wielded in the main hand; if in the off hand while
twoweaponing, it would still be possible to bribe the demon.

This commit adds a new function, wielding_artifact, that checks both the
main and off hand weapons for a particular artifact, and uses it for
most cases where artifact effects are activated.