New weapon: dwarvish bearded axe.
Elves and orcs get all these cool and diverse weapons, each even have
racial artifacts. Dwarves don't have much compared to the other two
races, so I've decided to spice things up a bit for them. The dwarvish
bearded axe: a one-handed long/bearded axe that deals out damage
comparable to a long sword (1d8 vs small, 1d10 vs large). Base material
is iron, but falls under the dwarvish weapons category for possible
materials purposes. Like all other axes, it's hefty, weighing in at 70
aum. Expect dwarves that you encounter in-game to potentially have one
of these as its primary weapon. If playing as a dwarvish valkyrie,
you'll start with this axe instead of a long sword.

Here's where it gets interesting. The design of a bearded axe made it so
it could be used to pull a weapon or shield out of a defender's grasp.
This axe has that same ability - the player just needs to apply the
bearded axe in the direction of their target, and they'll have a chance
of either disarming (if it's wielding a weapon) or removing the shield
(if it's wearing one) of their target. The odds are more favorable for
disarming versus removing a shield. The Valkyrie role is automatically
proficient at applying a bearded axe, just like an archeologist is
proficient at applying a bullwhip. Strength plays a factor in how well
you'll be able to successfully apply a bearded axe. Monsters can apply a
bearded axe against a player just the same as the player against
monsters.

Hopefully this adds a bit more flavor and fun to playing (and fighting
against) a dwarf.