Fix magic_detect monster detection
Reading a cursed scroll of magic detection would show monster glyphs,
but each one would be listed as "it" or "peaceful it" in farlook; this
fixes that issue.

I also moved the criteria for a monster being considered "magical" to a
macro so that it can be updated more easily in the future, given the
comment about that.