Kathryn's birthday.
If the day the player is playing happens to land on Kathryn's birthday,
certain things will change. Kathryn in both of her forms will have more
hit points, and she'll mention that it is her birthday when randomly
saying things to the player. I could do some more neat things with this,
but for now this sets the stage.

The extra functions added in src/hacklib.c I grabbed from UnNetHack.