Fix: prevent rangers from wishing for their other quest artifact.
Technically speaking, rangers have two quest artifacts - The Longbow of
Diana and The Crossbow of Carl. A gnomish ranger could wish for the
longbow, and a non-gnomish ranger could wish for the crossbow. This
could lead to bad game-breaking things happening with the quest leader
showing up as the owner of said artifacts.