Fix: random monster cannot spawn if unaligned (A_NONE).
Glad I caught this. PM_INFIDEL had it's alignment set as A_NONE which
wasn't an issue before since there used to be separate templates for
player monsters. Now that it's back to the way it should be, a monster
with an alignment of A_NONE causes a panic and crash if part of the
random monster generation code (currently, the only monster template
with an alignment of A_NONE is the Wizard of Yendor). So, set the
template for PM_INFIDEL to whatever, and then set an extra condition
during player monster generation that sets their alignment to unaligned
after the fact.

Also made a tiny unrelated edit to sysconf.