Fix: game freeze with rivers spawning in certain versions of minetown.
There are a couple versions of minetown that allow rivers to form. With
the recent introduction of commit 07df5eb, when those versions of
minetown would load, the game would freeze and it was stuck in an
infinite loop on making the map. Quick fix, just exclude minetown from
where the rivers can form. Had to make a new define for this, kinda
surprised it didn't exist already. This commit will also address the
rare moments that players would find jellyfish or piranha on dry land in
minetown.