Some minor adjustments to certain monster grudges.
If a huge or greater sized monster is going to stomp on tiny creatures
between the player and them to get to the player, they have to be fairly
close to the player now. Adjust the minimum hit points an M3_BERSERK
flagged monster has to reach before they lose it and just attacking
everything (33% to 20%).