Ensure magic traps ignore (see) invis timeouts
When the character has the invisible or see invisible properties from a
source that times out, like a potion, leave that be and toggle the
underlying permanent intrinsic without affecting the timeout.

For example, a character with see invisible only for another 30 turns
who steps onto a magic trap will have intrinsic see invisible enabled,
instead of having their see invisible timeout set to 0 and disabled.
This is more consistent with the way other external, temporary sources
of these attributes are treated by the magic trap toggle already.