Fix: 'The skeleton's skin looks flaky'.
There were a couple places in mcastu.c where if a monster had 'destroy
armor' spell cast on it and it wasn't wearing armor, and if it didn't
technically have skin, the feedback given was incorrect. I made a new
body_part() called SKIN, and it doesn't actually apply to this as of
right now (grammar issue - 'your skin itches' is fine, but not 'your
scales itches'). May tackle that later. There were a couple other places
though where body_part(SKIN) works.