Racial soldiers and other mercenary types.
With the advent of racial player monsters being possible and implemented
correctly, it now makes it relatively simple to do the same for
virtually any other type of monster as well. In this commit, racial
soldiers are implemented correctly, and other mercenary types (watchmen,
watch captains, guards, and prison guards) are included as well.