Cerberus can be put to sleep with a magic flute.
A bit of influence from Greek mythology - even though Cerberus is sleep
resistant, if the player plays a magic flute near him, he will fall
asleep for a short time. Not nearly as long a time as other monsters,
but long enough to buy the player enough time to flee or perform some
other time-intensive task.

Co-authored-by: entrez <me@entrez.cc>