Rename rndrace to m_randrace
The random race selection function for monsters was called 'rndrace',
which potentially could cause confusion since there was an existing
function for randomly selecting the player race called 'randrace' -- the
two are similar enough that I think it makes sense to change the name of
the monster function to be a bit clearer.

I also changed some of the mnum stuff from int to short, since that's
what is used for mnum in the monst struct.