Proper handling of monsters killed via artifact special fatal attacks.
Whenever a monster was killed from certain artifacts special attacks,
some of those monsters would be completely obliterated and not leave a
corpse. Was using mongone() for this, technically that means the monster
disappeared, not died. So no livelogging processed if that monster
happened to be unique, plus the player would not gain experience nor
credit for the kill. Lets swap mongone() for xkilled() instead and use
the proper settings to suppress leaving a corpse and an extra death
message.