Fix: force monster rider from steed if under influence of conflict.
Known issue for a while. If a monster riding a steed (or its steed) were
affected by conflict and started attacking each other and the steed
died, the rider would not be placed correctly on the map and would
appear to be invisible and also be impervious to harm. This commit
forces the monster rider and steed to separate if under the influence of
conflict once the steed dies, just like it would for the player and
its steed.

Co-authored-by: entrez <me@entrez.cc>