Prevent cannibalism from illithid tentacle attack against werecreatures if infected with lycanthropy.
If playing as illithid race, and the player became infected with
lycanthropy, if a successful tentacle attack was made against a
werecreature in its non-human form, it would count as cannibalism. Same
if the same attack landed on monsters associated with that wereform.
This commit suppresses the illithid's tentacle attack if they are
infected with lycanthopy. It's a bit of a kludge, would have rather used
were_beasties() instead of blocking certain mlet classes, but this will
get the job done for now.