Prevent water inhalation with magical breathing
Removing one source of magical breathing while underwater and still
wearing another, e.g. wearing sea dragon scale mail and an amulet of
magical breathing and removing the mail, would still cause a message
about "inhaling an unhealthy amount of water" to be printed as though
you had lost all sources of magical breathing.  This is beacuse the
amulet of magical breathing was previously only the source of this
property, so the code was not set up to check for other potential
sources.

This wouldn't happen when wearing the mail and amulet and removing the
amulet, but only because the mail also provides the "swimming" property
which prevents the message from being printed.  If the armor didn't
provide that property alongside magical breathing it would have happened
in that case as well.