Proper timeout for sphere spells.
Thanks AntiGulp (SpliceHack) for the point in the right direction. I
looked at how 'mfading' worked with that variant and applied it to
summoned spheres here. They now have a proper timeout, and the players'
skill level in casting matter-based spells plays a role. Summoned
spheres will last 100-200 turns at skilled or higher, or 50-100 turns at
basic or lower. I also improved the AI for these, so they attack pretty
much anything regardless of the level spread or type of monster. I also
switched 'isspell' to 'msummoned', makes a lot more sense, especially
for those viewing the code for the first time and wondering what it's for.