Demon boss/lair adjustments, new object.
* Various adjustments to several demon lords and princes (attack damage,
type, weapons they'll spawn with, etc)
* A new object has been introduced - the rod. A rod appears as an
'ornate mace' if not ID'ed, and is part of the mace family for
skill/damage type purposes. Damage output is 1d8 small/large. It will
not spawn randomly. This was created specifically for various demon
bosses, such as Asmodeus, Dispater, Orcus, etc. If you look up their
descriptions, they are always carrying a rod of some sort, usually made
of some type of gemstone or precious metal. Why fake the funk with a
mace? Let's just make the actual object.
* All demon lair levels are flagged 'hardfloor', exception being 'Orcus
town'
* To help make sure that there will always be a demon prince
(tier 3) present in its lair, an adjustment as been made to demon
summoning. Demon princes will never be able to be gated in until the
player has set foot on a tier 3 demon prince lair. Until then, demons
will only be able to gate in demon lords or other regular demons. Given
the increase in the number of new demon lords, this shouldn't pose any
kind of issue, and if the player wants to go for killing every demon
lord/prince that exists, they'll still be able to make it happen.

More to do still - new demon lair levels, some cool things to do with
Tiamat.