Lolth has two arms.
I forgot that Lolth isn't just a giant spider - she's more like a
drider, a humanoid upper torso on top of a (in this case, massive)
spider body. Meaning she has humanoid arms and hands and can use
weapons. Discovered that her favored weapon tends to be a scimitar, so
we'll set her up with one.