Set mspec for several spellcasting boss-type monsters to 0.
Something I noticed when play-testing the new demon bosses I've created;
many of them have two different spell attacks, but the second one would
*never* occur, except maybe from a distance if the first spell attack
couldn't be used at range. It's because mspec would never drop below 2,
which was effectively nerfing any monster with a second spell attack
(this nerf is present in vanilla NetHack as well and affects those
monsters with two spell attacks). This commit un-nerfs those monsters'
second spell attack. This will probably have a sigificant bearing on how
much more dangerous these monsters will now become. Which monsters are
un-nerfed now? Demon lords/princes, the Oracle, high priests, Kathryn
the Ice Queen (and Enchantress), quest leaders and nemesis, and the
Wizard of Yendor.

This is going to require some play-testing for sure. Perhaps set it so
mspec is 0 only some of the time if this proves to be too powerful.
We'll see.