New branch: Vecna's domain (part one).
Starting a new optional side branch/quest, featuring Vecna as the boss
monster to defeat at the end of the quest. Vecna is an uber-powerful
lich, and his history goes all the way back to Dungeons & Dragons early
days. In this setting, Vecna is the source of power for all other liches
existience. Liches can be destroyed in battle, but while Vecna exists,
their kind cannot be wiped out via genocide magic. Vecna must be
destroyed so the remaining scourge that spawned from him can be
destroyed once and for all.

This commit focuses on creating Vecna himself and developing the code
that prevents liches from being genocided until Vecna is destroyed.
Defeating Vecna will also count as an achievement, much like killing
Medusa or defeating the Ice Queen. Vecna is very powerful, and can cast
both wizard and clerical based monster spells. There's more to to with
Vecna still, but the foundation is made at least.

To do still - designing the branch levels, tweaking Vecna's stats and
adding gear, creating one of the famous artifacts that is associated
with Vecna and how the player could obtain it (the eye or the hand,
haven't decided which yet), and possibly introducing Vecna's lieutenant
Kas (who was also responsible for Vecna's eventual downfall).