New branch: Vecna's domain (part two) - artifacts, Kas (unique monster).
Part two of the 'Vecna's domain' commits. This one adds two new
artifacts - The Eye of Vecna and The Sword of Kas. I've also added the
monster template for Kas (Vecna's top lieutenant) and tweaked Vecna
himself a bit. Here's the details:

* The Eye of Vecna - an extremely powerful artifact that can only be
obtained if Vecna is killed in a certain way (not implemented in this
commit). The base item is an eyeball (new food object). The Eye of Vecna
confers cold resistance, ESP, astral vision, and half spell damage when
carried. When invoked, it will greatly reduce the hit points of all
living creatures that the player can or could see. However, this comes
at a significant cost to the player's luck and alignment record each
time it is invoked. If invoked when the player's luck is extremely low,
the Eye will turn on the player, instantly killing them. The Eye of
Vecna can be eaten, but this is not recommended, as Bad Things will
happen to the one that eats it. It can also be sacrificed at an altar
for max value, but this then permanently destroys the artifact.
* The Sword of Kas - this is the great sword (two-handed sword) forged
by Vecna himself for his top lieutenant, Kas. I repurposed the Sword of
Bheleu for this... sad to see it go, but it's for a good cause. The
Sword of Kas has many of the same attributes: it's chaotic, is
permanently poisoned, confers stoning resistance when wielded, and does
extra damage to non-aligned monsters. The extras, most of which I pulled
from ad&d lore - the sword confers giant-like strength when wielded,
it's bloodthirsty (same as Stormbringer), and if used against Vecna,
will dish out triple the normal damage.
* Neither of these new artifacts can be wished for.
* Kas - the right-hand man of Vecna. Was legendary in batle, rumored to
always come out unscathed. Kas is actually a fairly powerful vampire,
second only to Vlad. Per legend, Kas betrayed Vecna and destroyed him,
producing the Eye and the Hand of Vecna. Perhaps the player can beat him
to it...

Next commit will be development of how the Eye of Vecna is obtained, as
well as gear outfitting for both Vecna and Kas.