Kit out Vecna and Kas, other tweaks and adjustments.
Give Vecna and Kas their gear. Vecna will spawn with a magical cursed
robe, and there's a 25% chance he will death drop the Eye of Vecna when
destroyed. Kas is all decked out in armor, and wields the Sword of Kas.

Other tweaks:
* Better feedback when trying to geno L or alhoons while Vecna exists
* Changed color of Vecna back to bright magenta
* Added some rumors (true an false) surrounding Vecna
* The Sword of Kas is excluded from the number of artifacts in-game when
determining chances for an artifact wish, as it's guaranteed to appear