New branch: Vecna's domain (part three) - levels, minor tweaks.
This is the third and final major commit to finish out the new side
branch featuring Vecna. This branch features three custom levels: the
initial landing level, the stronghold of Kas, and finally Vecna's
domain. Each level will be harder than the one before it. Some minor
tweaks were made here as well concerning Kas (he kept wanting to change
into a wolf or a bat, not good if he starts with inventory).

Expect more adjustments and tweaks in the future, but as of now, this
optional side quest is complete.