Intelligent monster bag preferences
Don't prevent a monster from picking up a bag just because the hero
doesn't have a bag of the exact same type -- this would prevent a
monster from picking up a sack if the hero has only a bag of holding.
Instead, pick up a bag if the hero has one of the same type or better,
and exchange worse bags for better ones.