Fix: segfault if killed by certain monsters that go through your possessions.
Recently had a player have their game crash after death but before the
dumplog and xlogfile could completely be written to. Case: monster that
is removed from bones becuase they have a home base they absolutely need
to be found in (that works) but then another part of the code expected
them to start going through leftover inventory (game crashes because
expected monster is gone). Didn't get a backtrace on this crash, but
logically, this seems to be what is happening.

Also did some code formatting cleanup here.