Abusing your alignment can cause your quest leader to ask you to return the quest artifact.
This is an idea I came up with a long time ago. You set out to start
your quest, meet up with your quest leader, and they go on to tell you
how the quest artifact was stolen, how important it is that you get it
back for them for the safety of everyone, so on and so forth. But then
when you do return it, the quest leader says 'oh hey, keep it, it's
yours'. That always seemed a bit weird to me. I always wondered 'what if
the quest leader tells you 'great, you got the quest artifact back, hand
it over...' What would you do? Keep it and piss off your quest leader?
Hand it over for the greater good with only the Bell of Opening to show
for it?

This commit explores those possible outcomes. Kudos to aosdict for
coming up with the idea for tying this in with negative alignment
changes. Here's the breakdown:

* If you abuse your alignment, that is now tracked (u.ualign.abuse)
* Your alignment abuse score mirrors that of any negative alignment
record adjustments - kick over a grave? That adversely adjusts your
alignment record by one, and also adjusts the abuse ticker by the same
amount. Kill your pet? That's 15 points for both. You get the idea.
* Unlike your alignment record, you can't reverse the amount of points
racked up via alignment abuse
* When you finish your quest and return to your quest leader with the
quest artifact, the game looks at how many alignment abuse points you've
accumulated. Having only one alignment abuse point will cause your quest
leader to possibly ask you to return the quest artifact - just one abuse
point puts those odds of being asked at 1 in 50. Additional points
increase those odds (2 points makes it 1 in 49 odds, 3 points 1 in 48,
etc)
* If you've never abused your alignment by the time you reach your quest
leader with the quest artifact, you'll never be asked to hand it over.
* Infidels will also never be asked for their quest artifact, as they
need it to win the game. I may change this in the future
* If your quest leader asks you to hand over the quest artifact, you'll
be asked if you want to hand it over (yes or no). Answering yes removes
the quest artifact from your inventory and puts in the quest leader's
inventory. The player gets a boost to their luck and alignment score,
and the quest is flagged as complete. Answering no, you keep the quest
artifact, but you take a hit to your alignment, and your quest leader
plus any remaining guardians get angry with you. The quest is flagged as
complete.
* To prevent the player from skipping out on talking to their quest
leader after obtaining the quest artifact, their quest is not flagged as
complete until they do so. The Bell of Opening will not work for the
invocation until the quest is flagged as complete - the story arc has
been tweaked a bit to inform the player that the 'silver bell has been
cursed' and that the player will need to bring the quest artifact to
their quest leader; with it, they can 'uncurse' the bell so that it will
function properly
* If you wind up killing your quest leader before starting your quest,
simply killing your quest nemesis afterwards will flag the quest as
complete
* Conversely, if you come back with the quest artifact and somehow make
your quest leader angry before you're able to talk to them to mark the
quest as complete, killing the quest leader will then flag the quest as
complete
* Alignment abuse is now also a conduct, which will show up in the
xlogfile as 'never_abused_alignment'
* You can also check to see if you've ever abused your alignment
via the #conduct command in-game - if you don't see anything about not
abusing your alignment, it means you've abused it at some point

This adds a whole new dimension to every quest, regardless of role
(Infidel as the only exception). Staying true to your alignment has now
become much more important, especially in the early/mid game. As this is
the initial commit, there may be more tweaks and adjustments to this in
the future. Thank you to everyone who provided feedback and constructive
criticism while I was creating this.