Fix: mon 457 doesn't like any materials for obj 269.
There's a bit in makemon.c that prevents monsters from spawning with
objects made of a material they hate. The panic message above was being
generated when entering the final archeologist level (quest nemesis is
technically a demon, and it handles the Bell of Opening which is made of
silver). Exclude the Bell from that check.