Randomize odds of a river forming in the gnomish mines.
Thought it would be neat to add some randomness to the gnomish mines.
Instead of a river (or sewer) forming on all regular mines levels all
the time, it's random. How often a river will form depends on how deep
into the mines the player goes (greater odds the deeper they descend).
I've also added a bit of randomness to what monsters can spawn in the
river - if the mines level depth is 9 or greater, there's a small chance
a water moccasin could spawn (or crocodile if sewers), and then an even
smaller chance of a water troll spawning instead of a water moccasin.
For river formation, overall the odds are 50-75% of seeing a river on
the very first mines level, and then it goes up from there.

This could be improved upon (see comments in mkmap.c).