Give the player a chance to escape from a demon lord/prince using teleportation or cursed potions of gain level.
Randomness and unexpected things happening are a big part of NetHack. I
figured this bit needed a bit of that, and to not be so strict. Instead
of never being able to teleport/cursed potion of gain level away from a
demon lord or prince, there's now a small chance it can work. With demon
lords that chance is 1 in 5. With demon princes, the odds go to 1 in
20.