Wizard quest leader/nemesis tweaks, MR2_TELEPATHY extrinsic for monsters from worn objects.
This initially started out as trying to get the Dark One to wear the Eye
of the Aethiopica. Before, just carrying the EotA conferred magic
resistance. Since 3.6.x, you now have to wear it to get MR. This in turn
nerfed the Dark One because the code didn't exist to make a monster want
to wear an amulet of ESP/the EotA, and he thereby lost his MR. This
commit addresses that - monsters will want to take and wear amulets of
ESP or the EotA. They'll get the ESP monster extrinsic from wearing it
(same with a helm of telepathy or an object with the object property).
They will also get MR from wearing the EotA (this is technically a
change that should have occurred in vanilla NetHack, but this goes back
to nothing wants to wear amulets of ESP). So after all of this, the
Dark One still won't wear it. Maybe it's because his alignment doesn't
match the EotA. Regardless, to rectify him being nerfed, I've kitted
him out with proper wizard attire, including a cloak of magic
resistance and a pointy hat. Neferet gets a robe and a pointy hat, too.
The Dark One has also been given an extra spellcasting attack so he's on
par ability-wise with Neferet. I was told earlier that the Dark One was
a pushover now since 3.6.x, so this should nip that problem in the bud.