Fix: various objects incapable of being ID'ed under certain conditions.
'If you wish for a fooproof item (or get it as a sac gift) that doesn't
have a unique unidentified description and then learn its BUC status,
there's no way to identify it as fooproof (e.g. wish for a "fixed
stethoscope" and drop it on an altar). You can't ID it further after
that, but if you wish for a "fixed stethoscope" and then read ID on it
before learning its BUC status, it will mark it as rustproof.'

Co-authored-by: entrez <me@entrez.cc>