Barbed devils use hellfire.
Swap out one of the barbed devil's physical attacks with the ability to
cast 'hellfire' at its target. This uses AT_MAGC/AD_FIRE, so it operates
just like an angel using shower of missiles if up close or just plain
magic missile from afar, but in this case its fire. If the target is
100% fire res, they can still take half-damage from hellfire if it's
used up close.