Fixed and inherently fixed rings/wands can resist destruction due to shock damage.
If you have reflection and are zapped with a wand of lightning or
something similar, your rings and wands in open inventory are protected
from shock damage. Not so if attacked by a monster wielding a <weapon>
of lightning, which does happen. Because of the odds of having your
rings and wands subject to shock damage has increased a fair amount, it
seemed fair to give said items a fighting chance. If your ring or wand
is fixed (fooproofed), or if it's already made out of a material that
is inherently fixed, it has a 66% chance of resisting being destroyed.