Barding.
This commit adds a new 'tool' type - barding for steeds. If you're
unfamiliar with what barding is, in this case it's armor for your pet
steed (and also for monster steeds). In medieval times, barding came
in all sorts of configurations, but for our purposes here, imagine it
protecting the steeds head, neck, chest, and its flank from its chest
back to the start of the saddle. It works much the same way you'd
put on/remove a saddle, and the types of monsters barding can be placed
on are virtually the same. Since saddles don't care about the size of
the steed, I decided that barding shouldn't either (that may be
something for another day). Barding can't be enchanted like normal
armor, its material dictates how much AC protection it provides. There
are three types of barding - regular barding, spiked barding and
barding of reflection (polished barding). Regular barding is just that.
Spiked barding has several spikes that protude outwards and will
prevent the steed wearing it from dying via engulf attack (ever watch
something try to eat a pufferfish live? Kinda like that). The last acts
just like a shield of reflection. If a monster rider/steed pair spawn,
and its steed has a saddle, there's a small chance the steed will also
be barded. All barding types offer protection against attacks to the
head, just like wearing a helmet, and it can be greased as well.

This was fun to make and test. More could probably be done here, but
this is a good start for now. This should improve steed survivability
significantly.