Koa tribute.
Our oldest dog Koa passed away late last year. It's nice that she is
immortalized in a way being in this game, but she's no longer with us.
A fitting tribute feels appropriate. We recently added another dog to
the family, her name is Bourbon and she's a bit younger than Ozzy, same
breed. She now takes Koa's spot on the final level, and Koa has a nice
gravesite and a fitting epitaph next to the Ice Queen's throne.