Gnolls.
This commit adds another humanoid monster to the game - Gnolls. These
creatures are common in Dungeons & Dragons: they are human-like, with
a strong resemblance to hyenas, and not just physically. In-game they
will always appear as a lowercase 'h' (colors stated below). Gnolls
are present in a couple other variants (notably Slash'EM and dNetHack),
but in those variants, gnolls are of high level and extremely powerful
and dangerous. For EvilHack, I've taken a more traditional route, the
types and their stats taken from ad&d and 'Volo's Guide to Monsters'.
Players will start to encounter gnolls towards the end of the early
game/midgame, as well as certains levels in Gehennom. Here are the
different types that can be encountered:

* Gnoll - base level 4 (brown), has one weapon and one bite attack. Can
appear in large groups
* Gnoll Hunter - base level 7 (orange). A stronger, faster gnoll that
hunts in small groups and will always have a bow and a number of arrows
* Gnoll Cleric - base level 8 (light blue), a physically weaker gnoll,
but can cast an array of monster-based clerical spells
* Flind - base level 10 (yellow). Certain gnolls that were born not just
stronger and faster, but smarter than the rest. Flinds tend to rule over large
groups of gnolls and will lead the larger, more organized raiding
parties
* Gnoll Witherling - base level 5 (orange). This is technically a gnoll
zombie, but in Dungeons & Dragons they are referred to as witherlings.
Any gnoll killed by a zombie will rise again as a gnoll witherling

All gnolls (except clerics) spawn with some sort of weapon, typically a
flail, spear, morning star, or polearm, while flinds can also have a
battle-axe or two-handed sword. Gnolls will also have some sort of
leather or cloth based armor for defense.

Yeenoghu will always have an entourage of various gnolls near him when
encountered in his lair.