Giant spiders, Lolth can entangle their target with a web attack.
Took some inspiration from dNetHack, created a new 'damage' type that on
a successful hit, entangles the target in a web trap. If the
target is already in a trap, the web is not created. Currently only
giant spiders and Lolth have this attack via AT_TUCH (zero damage), and
giant spiders have had their difficultly adjusted up because of it.

Monsters that would normally automatically tear through a web do the
same versus this form of attack. Monsters that can already create webs
are also unaffected. If the target is on a steed, being entangled causes
the target to fall off its steed.

Because this new damage type makes giant spiders more deadly, I adjusted
the minimum depth of the 'spider room' vault to dlvl 5 (was dlvl 3).

In the future I may attach this damage type to AT_SPIT attack (imagine
shooting webs at a monster from range, like Spider-Man.